+++ changeLog.md
@@ -1,6 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.3.0] - 2018-07-01
+### Added
+- ChannelId in contructor
+### Changed
+- GetViewers now call a url for get the real viewers
+- GetChannel now set the channelId and return information
+- LiveId change position of contructor
+- liveBroadcast method now call `search` of uri Api
+### Removed
+- ChatId in params of constructor
+
 ## [1.2.0] - 2017-10-18
 ### Added
 - ChatId in params of constructor